- changed library dependency dir to lib_precompiled/include/ in msvc properties - fixed a bug with tolua pkg file in msvc - (u)int64_t to (unsigned) long long (C-Standard, so most c++ will have it anyway)

+++ code/branches/gui/src/core/Clock.cc
@@ -71,7 +71,7 @@ namespace orxonox
         }
     }
 
-    uint64_t Clock::getRealMicroseconds() const
+    unsigned long long Clock::getRealMicroseconds() const
     {
         return this->timer_->getMicroseconds();
     }
